Create basketball-game.py
At the beginning of the game, you start with an empty record. You are given a list of strings ops, where ops[i] is the ith operation you must apply to the record and is one of the following:

An integer x - Record a new score of x,

"+" - Record a new score that is the sum of the previous scores. It is guaranteed there will always be two previous scores.

"D" - Record a new score that is double the previous score. It is guaranteed there will always be a previous score.

"C" - Invalidate the pevious score, removing it from the record. It is guaranteed there will always be a previous score.

Return the sum of all the scores on the record.